Organize project setup, scalac settings

+++ src/main/scala/faker/time/FutureInstant.scala
@@ -12,6 +12,8 @@ object FutureInstant {
       now <- Arbitrary.arbitrary[NowInstant].map(_.value.plusMillis(1L))
       epochSecond <- Gen.choose(now.getEpochSecond, Instant.MAX.getEpochSecond)
       nanoAdjustment <- Gen.choose(now.getNano, Instant.MAX.getNano)
-    } yield FutureInstant(Instant.ofEpochSecond(epochSecond, nanoAdjustment))
+    } yield FutureInstant(
+      Instant.ofEpochSecond(epochSecond, nanoAdjustment.toLong)
+    )
   }
 }
